\docType{package}
\name{ALKr}
\alias{ALKr}
\alias{ALKr-package}
\title{Calculating Age-Length Keys from incomplete data}
\description{
  Generate Age-Length Keys for fish populations from
  complete or incomplete data.
}
\details{
  This will be a more detailed description.
}
\references{
  Alberto Murta, Jose Francisco Loff, Laura Wise, Manuela
  Neves, Laurence Kell.
}

