\docType{package}
\name{ALKr}
\alias{ALKr}
\alias{ALKr-package}
\title{Calculating Age-Length Keys from incomplete data}
\description{
  Generate Age-Length Keys for fish populations from
  complete or incomplete data.
}
\details{
  A collection of functions that implement several
  algorithms for generating age-length keys for fish
  populations from incomplete data.
}
\references{
  Alberto Murta, Jose Francisco Loff, Laura Wise, Manuela
  Neves, Laurence Kell.
}

